Split multimodal graph tests by modality
Separate vision-language coverage from speech, ASR, TTS, and diarization tests so each L1 module has a cohesive scope. Update all remaining stale references to deleted graph-test modules.
